fix(web): address app shell store code review feedback
Stabilize navigation custom hook deps with refs and primitive
signatures, use useState for store init, move sidebar hook to its own
file, and stop re-syncing sidebar on provider re-renders.
